Move tests into a separate directory
Also run these tests on Github Actions as part of the PR request
flow.
