+++ tun/src/unix/linux/mod.rs
@@ -172,32 +172,3 @@ impl TunInterface {
         self.socket.send(buf)?
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::TunInterface;
-    use std::net::Ipv4Addr;
-
-    #[test]
-    fn mtu() {
-        let interf = TunInterface::new().unwrap();
-
-        interf.set_mtu(500).unwrap();
-
-        assert_eq!(interf.mtu().unwrap(), 500);
-    }
-
-    #[test]
-    #[throws]
-    fn netmask() {
-        let interf = TunInterface::new()?;
-
-        let netmask = Ipv4Addr::new(255, 0, 0, 0);
-        let addr = Ipv4Addr::new(192, 168, 1, 1);
-
-        interf.set_ipv4_addr(addr)?;
-        interf.set_netmask(netmask)?;
-
-        assert_eq!(interf.netmask()?, netmask);
-    }
-}
